Sall-eGarbageDetection/requirements.txt
# Image processing
opencv-python-headless
pillow-avif-plugin
pillow
numpy

# Video processing
ffmpeg-python
torchvision
torchaudio
timm

# Model setup
ultralytics
transformers
torch
git-lfs
yolov5

# Server startup
uvicorn
fastapi
python-multipart
huggingface_hub==0.14.1